Update all contexts to encapsulate specific properties
