+++ src/App.jsx
@@ -1,5 +1,4 @@
 import "./App.css";
-import { useGlobalContext } from "./context";
 import Header from "./components/Header";
 import Footer from "./components/Footer";
 import Alerts from "./components/Alerts";
@@ -9,13 +8,14 @@ import NavOptions from "./components/NavOptions";
 import Filters from "./components/Filters";
 import MapContainer from "./components/MapContainer";
 import { useAlertContext } from "./AlertContext";
+import { useLoginContext } from "./LoginContext";
 
 // I placed test text and context placeholder text for
 // everyone to hopefully understand how the context works.
 // It will show up in the test text main, and also the developer console.
 
 function App() {
-  const { login } = useGlobalContext();
+  const { login } = useLoginContext();
   const { showModal } = useAlertContext();
 
   return (
