+++ src/components/Modal.jsx
@@ -1,4 +1,4 @@
-import { useGlobalContext } from "../context";
+import { useAlertContext } from "../AlertContext";
 
 const Modal = () => {
   const { selectedAlert, closeModal, modalImage } = useAlertContext();
