+++ src/index.jsx
@@ -2,18 +2,21 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { AlertProvider } from "./AlertContext";
 import App from "./App";
-import { AppProvider } from "./context";
+import { NavigationProvider } from "./NavigationContext";
+import { LoginProvider } from "./LoginContext";
 import { PolygonProvider } from "./PolygonContext";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
-    <PolygonProvider>
-      <AppProvider>
+    <LoginProvider>
+      <NavigationProvider>
         <AlertProvider>
-          <App />
+          <PolygonProvider>
+            <App />
+          </PolygonProvider>
         </AlertProvider>
-      </AppProvider>
-    </PolygonProvider>
+      </NavigationProvider>
+    </LoginProvider>
   </React.StrictMode>
 );
